% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_candidates_2020.R
\docType{data}
\name{act_candidates_2020}
\alias{act_candidates_2020}
\title{List of candidates in the 2020 ACT election.}
\format{
A data frame with 137 rows and 5 variables:
\describe{
  \item{party}{Party name}
  \item{electorate}{Electorate name}
  \item{last_name}{Candidate's last name}
  \item{first_name}{Candidate's first name}
  \item{ballot_paper_name}{Name as it would appear on the ballot paper, for joining to \link{act_fp_2020}}
}
}
\source{
\url{https://www.elections.act.gov.au/elections_and_voting/2020_legislative_assembly_election/list-of-candidates}
}
\usage{
act_candidates_2020
}
\description{
Data is copyright to the ACT Electoral Commission
}
\keyword{datasets}
