% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_fp_2019.R
\docType{data}
\name{act_fp_2019}
\alias{act_fp_2019}
\title{First preference results for the 2019 ACT election.}
\format{
A data frame with 13289 rows and 3 variables:
\describe{
  \item{candidate}{Candidate's name}
  \item{votes}{Number of formal votes}
  \item{polling_place}{Name of the polling place}
  ...
}
}
\source{
\url{https://www.electionresults.act.gov.au/Results/PollingPlace}
}
\usage{
act_fp_2019
}
\description{
Data is copyright to the ACT Electoral Commission
}
\keyword{datasets}
