% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_polling_places_2020.R
\docType{data}
\name{act_polling_places_2020}
\alias{act_polling_places_2020}
\title{Polling places for the 2020 ACT election}
\format{
An object of class \code{sf} with 82 rows and 3 columns.
}
\source{
\url{https://www.elections.act.gov.au/elections_and_voting/2020_legislative_assembly_election/polling-places-2020}
}
\usage{
act_polling_places_2020
}
\description{
Data is copyright to the ACT Electoral Commission
}
\keyword{datasets}
