% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act_preferences_2019.R
\docType{data}
\name{act_preferences_2019}
\alias{act_preferences_2019}
\title{Preference distribution for the 2019 ACT election}
\format{
A data frame with 13289 rows and 3 variables:
\describe{
  \item{count}{Stage of the count}
  \item{type}{Type of result, either distribution or final}
  \item{name}{Candidate's name}
  \item{votes}{Number of votes}
  \item{electorate}{Electorate name}
}
}
\source{
\url{https://www.elections.act.gov.au/elections_and_voting/2020_legislative_assembly_election/distribution-of-preferences-2020}
}
\usage{
act_preferences_2019
}
\description{
Data is copyright to the ACT Electoral Commission
}
\keyword{datasets}
